Refactored application package to com.google.wireless.debugging.app, added wireless debugging module
