+++ mobile/android/WirelessDebug/wirelessdebugging/.gitignore
@@ -1,0 +1,1 @@
+/build
