+++ mobile/android/WirelessDebug/wirelessdebugging/src/main/java/com/google/wireless/debugging/WirelessDebugger.java
@@ -1,0 +1,9 @@
+package com.google.wireless.debugging;
+
+
+public class WirelessDebugger {
+
+
+
+
+}
